Split tab control into its own component
